+++ pubspec.yaml
@@ -4,7 +4,8 @@ version: 1.0.0+1
 publish_to: none
 
 environment:
-  sdk: ">=3.0.0 <4.0.0"
+  sdk: ">=3.8.0 <4.0.0"
+  flutter: ">=3.8.0"
 
 dependencies:
   auto_route: ^10.1.2
@@ -38,6 +39,8 @@ dependencies:
   json_annotation: ^4.9.0
   retrofit: ^4.7.2
   dotenv: ^4.2.0
+  shadcn_ui: ^0.31.8
+  lucide_icons_flutter: ^3.1.0
 
 dev_dependencies:
   flutter_test:
